Adapt for using ScanImageTiffReader

+++ iCorre_internal/iCorre_batch.m
@@ -14,6 +14,11 @@
 %   logical status:     Logical mask for successfully processed data directories.
 %   cell msg:           Any associated error messages, indexed according to data directory.
 %
+%NOTES
+%   --  Occasionally, writing to a working MAT file fails during
+%       registration, throwing this error: 'MATLAB:save:permissionDenied'
+%       If problem persists, try checking attributes and pausing
+%       before save...
 %-----------------------------------------------------------------------------------------------------------
 
 function [ status, err_msg ] = iCorre_batch(root_dir,search_filter,params)
@@ -27,6 +32,10 @@ if nargin<3
         search_filter = '';
     end
 end
+
+%% Check Inputs
+params = iCorreCheckParams(params); %**Work-in-progress**
+
 
 %% Get list of data directories
 temp = dir(fullfile(root_dir,search_filter)); %Edit to specify data directories
@@ -53,9 +62,10 @@ for i=1:numel(data_dirs)
         dirs.main = fullfile(root_dir,data_dirs{i});
         dirs.raw = fullfile(root_dir,data_dirs{i},'raw');
         dirs.mat = fullfile(root_dir,data_dirs{i},'mat');
-        dirs.save = fullfile(root_dir,data_dirs{i},'registered'); %to save registered stacks as TIFF
-        if params.ref_channel~=params.reg_channel && params.do_stitch %If two-color co-registration
-            dirs.save_ref = fullfile(root_dir,data_dirs{i},'registered_ref_channel'); %Save dir for registered reference channel
+        dirs.save_mat = fullfile(root_dir,data_dirs{i},'registered mat');
+        dirs.save_tiff = fullfile(root_dir,data_dirs{i},'registered tiff'); %to save registered stacks as TIFF
+        if params.ref_channel && params.reg_channel && params.do_stitch %If two-color co-registration
+            dirs.save_ref = fullfile(root_dir,data_dirs{i},'registered_ref_channel'); %Save dir for registered reference channel (for two-color co-registration)
         end
         
         % If No 'raw' Directory, Create It & Move the TIF Files There
@@ -63,12 +73,19 @@ for i=1:numel(data_dirs)
             mkdir(dirs.raw);
             movefile(fullfile(dirs.main,'*.tif'),dirs.raw);
         end
+
+        % Delete any existing MAT directory
+        if exist(dirs.mat,'dir')
+            rmdir(dirs.mat,'s'); %MAT files for saving registration-in-progress
+        end
+
         % Create remaining directories
         field_names = fieldnames(dirs);
         for j=1:numel(field_names)
-            if ~exist(dirs.(field_names{j}),'dir')
-                mkdir(dirs.(field_names{j}));
-            end
+%             if ~exist(dirs.(field_names{j}),'dir')
+%                 mkdir(dirs.(field_names{j}));
+%             end
+            create_dirs(dirs.(field_names{j}));
         end
         
         % Define All Filepaths Based on Data Filename
@@ -76,8 +93,8 @@ for i=1:numel(data_dirs)
         file_names = cell(numel(temp),1);
         for j=1:numel(temp)
             file_names{j} = temp(j).name;
-            paths.raw{j} = fullfile(dirs.raw,file_names{j});
-            paths.mat{j} = fullfile(dirs.mat,[file_names{j}(1:end-4) '.mat']); %for .mat file
+            paths.raw{j} = fullfile(dirs.raw,file_names{j}); %Raw TIFFs for registration
+            paths.mat{j} = fullfile(dirs.mat,[file_names{j}(1:end-4) '.mat']); %MAT file (working file for read/write across iterations)
         end
         
         % Additional Paths for Metadata
@@ -93,27 +110,45 @@ for i=1:numel(data_dirs)
         
         %% Load raw TIFs and convert to MAT for further processing
         disp('Converting *.TIF files to *.MAT for movement correction...');
-        stackInfo = get_imgInfo(paths.raw, params); %Extract header info from image stack (written by ScanImage)
-        stackInfo.tags = tiff2mat(paths.raw, paths.mat, params.ref_channel); %Batch convert all TIF stacks to MAT and get info.
+        stackInfo = tiff2mat(paths.raw, paths.mat,...
+            struct(...
+            'chan_number',params.ref_channel,...
+            'crop_margins',params.crop_margins,...
+            'extract_I2C',params.saveI2CData)); %Batch convert all TIF stacks to MAT and get info.
         save(paths.stackInfo,'-STRUCT','stackInfo','-v7.3');
         
         %% Correct RIGID, then NON-RIGID movement artifacts iteratively
         % Set parameters
         RMC_shift = round(0.5*params.max_shift); %Params value used for SEED (based on reference image); therefter, narrow the degrees freedom
         NRMC_shift = round(0.5*params.max_shift); %[10,10,0]; max dev = [5,5,0] for 256x256
-        overlap = params.grid_size/4; %[16,16] for 256x256
+        overlap = round(params.grid_size/4); %[16,16] for 256x256
         
         options.seed = NoRMCorreSetParms('d1',stackInfo.imageHeight,'d2',stackInfo.imageWidth,...
-            'max_shift',params.max_shift,'shifts_method','FFT',...
-            'boundary','zero','upd_template',false,'use_parallel',true,'output_type','mat'); %Initial rigid correct for drift
+            'max_shift',params.max_shift,'shifts_method','FFT',... 
+            'correct_bidir',true,'bidir_us',1,... %correct_bidir appears to introduce artifacts when us_fac>1; use with caution.
+            'boundary','NaN','upd_template',false,... 
+            'print_msg',false); %Initial rigid correct for drift
         options.RMC = NoRMCorreSetParms('d1',stackInfo.imageHeight,'d2',stackInfo.imageWidth,...
             'max_shift',RMC_shift,'shifts_method','FFT',...
-            'boundary','zero','upd_template',false,'use_parallel',true,'output_type','mat'); %Rigid Correct; avg whole stack for template
+            'correct_bidir',false,...  %use correct_bidir only on seed
+            'boundary','NaN','upd_template',false,...
+            'print_msg',false); %Rigid Correct; avg whole stack for template
         options.NRMC = NoRMCorreSetParms('d1',stackInfo.imageHeight,'d2',stackInfo.imageWidth,......
             'grid_size',params.grid_size,'overlap_pre',overlap,'overlap_post',overlap,... 
-            'max_shift',NRMC_shift,'max_dev',params.max_dev,'correct_bidir',false,'shifts_method','FFT',... %Only cubic shifts are supported using normcorre_batch_even 
-            'boundary','zero','upd_template',false,'use_parallel',true,'output_type','mat'); %Non-rigid Correct; {'correct_bidir',true} threw error in iCorre.m on some data sets.
+            'max_shift',NRMC_shift,'max_dev',params.max_dev,...
+            'correct_bidir',false,...
+            'shifts_method','cubic',... %Only cubic shifts are supported if using normcorre_batch_even 
+            'boundary','NaN','upd_template',false,...
+            'use_parallel',true,'print_msg',false); 
+
+        %--- Notes on NoRMCorreSetParms ---
+        %'use_windowing' not supported when providing a template image.
+        %----------------------------------
         
+        %Check params
+        params.imageSize = [stackInfo.imageHeight, stackInfo.imageWidth];
+        params = iCorreCheckParams(params);
+
         % Initialize .mat file
         save(paths.regData,'params','-v7.3'); %so that later saves in the loop can use -append
         
@@ -126,9 +161,10 @@ for i=1:numel(data_dirs)
             disp(' ');
             disp([upper(options_label{m}) ' registration in progress...']);
             
-            [template,nReps,err_mat.(options_label{m})] = ...
+            
+            [template,nReps] = ...
                 iCorre(paths.mat,options.(options_label{m}),options_label{m},template,...
-                params.max_err,params.max_reps(m));
+                params.max_err,params.max_reps(m)); 
             
             template_out.(options_label{m}) = template;
             nRepeats.(options_label{m}) = nReps;
@@ -148,27 +184,25 @@ for i=1:numel(data_dirs)
         tic;
         
         %Apply registration to second channel if needed
-        if params.ref_channel~=params.reg_channel
-            paths.save = applyShifts_batch(paths,dirs.save,stackInfo,params.reg_channel); %Apply shifts and save .TIF files
+        params.fileType_save = ["tiff","mat"]; %***for testing; later, add to params***
+        if params.ref_channel && params.reg_channel %set to 0 for 1-channel imaging
+            paths = applyShifts_batch(paths,dirs,stackInfo,params.reg_channel,params); %Apply shifts and save .TIF files
             if params.do_stitch
-                disp('Getting global downsampled stack (binned avg.) and max projection from reference channel...');
+                disp('Getting global downsampled stack (binned avg.) and max projection from registered reference channel...');
                 binnedAvg_batch(paths.mat,dirs.save_ref,stackInfo,params.bin_width); %Save binned avg and projection to ref-channel dir
                 if params.delete_mat
                         rmdir(dirs.mat,'s'); %DELETE .MAT dir...
                 end
                 disp('Getting global downsampled stack (binned avg.) and max projection of co-registered frames...');
-                binnedAvg_batch(paths.save,dirs.main,stackInfo,params.bin_width); %Save binned avg and projection to main data dir
+                binnedAvg_batch(paths.save_tiff,dirs.main,stackInfo,params.bin_width); %Save binned avg and projection to main data dir
             end
             
         else
             %Save registered stacks as .TIF
-            for k = 1:numel(file_names)
-                S = load(paths.mat{k},'stack'); %load into struct to avoid eval(stack_names{k})
-                saveTiff(S.stack,stackInfo.tags,fullfile(dirs.save,[options_label{m} '_' file_names{k}])); %saveTiff(stack,img_info,save_path))
-            end
+            paths  = applyShifts_batch(paths,dirs,stackInfo,[],params); %Apply shifts and save .TIF files
             if params.do_stitch %Generate global and summary stacks for quality control
                 disp('Getting global downsampled stack (binned avg.) and max projection of registered frames...');
-                binnedAvg_batch(paths.mat,dirs.main,stackInfo,params.bin_width); %Save binned avg and projection to main data dir
+                binnedAvg_batch(paths.save_tiff,dirs.main,stackInfo,params.bin_width); %Save binned avg and projection to main data dir
             end
             if params.delete_mat
                 rmdir(dirs.mat,'s'); %DELETE .MAT dir...
@@ -184,7 +218,7 @@ for i=1:numel(data_dirs)
         
     catch err
         disp(err);
-        disp([{err.stack.file}' {err.stack.line}']);
+        disp([{err.stack.name}',{err.stack.line}']);
     end %end try
     
     if ~exist('err','var')
@@ -198,4 +232,5 @@ for i=1:numel(data_dirs)
     
     clearvars '-except' root_dir data_dirs params i status err_msg;
 end %end <<for i=1:numel(data_dirs)>>
+
